+++ dpws/src/main/java/org/somda/sdc/dpws/http/helper/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * HTTP server and client helper classes.
+ */
+@ParametersAreNonnullByDefault
+package org.somda.sdc.dpws.http.helper;
+
+import javax.annotation.ParametersAreNonnullByDefault;